POC-Snowflake/weather_pipeline_project/requirements.txt
pandas
snowflake-connector-python
python-dotenv
dash
plotly
